github: updated WF's to use paths-ignore
Updated the workflow configuration to include paths-ignore for Markdown
file changes. This adjustment ensures that the build and test processes
are skipped when only Markdown files are modified.

Refactored the get_run_id function to utilize the GitHub API instead of
relying on the previous GitHub script. This change improves the
reliability and efficiency of retrieving the run ID by
leveraging GitHub's API directly

Added a condition to trigger the `check_md_files` job only when the PR
is approved.

Updated check_md_files to include a check for build.yml. This ensures
that the job checks for changes to the build.yml file, preventing the
eden.yml workflow from failing if the PR contains changes exclusively to
build.yml.

Signed-off-by: yash-zededa <yash@zededa.com>