mini-yetus : Improve error reporting
Improve error reporting in mini-yetus.sh script and don't exit if files
dosn't exist. Instead, print an error message and continue. This can happen
if the current branch is built on top of the source branch but source
branch has diverged and some files are not present in the current branch.
This changes allow the script to atleast run yetus on the files that do
exist in the current branch.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>