memory-monitor:  Move CHECK_INTERVAL_SEC constant to shared config.
The `CHECK_INTERVAL_SEC` constant, previously defined in `procfs.c`, has
been moved to `config.h` for better maintainability and reusability.
This change centralizes the constant definition, as it will be used for
setting up a new monitoring thread in future updates. By placing it in
the shared config, we ensure consistency across different components
that rely on the same interval.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>