flextimer/flextimer.go: assert on 0 minTime for the exponential backoff
Zero `minTime` parameter does not make any sense for the exponential
backoff, because in case of repeated wakup `minTime` has to be doubled.
This patch adds an assert for the incorrect parameter, luckly we have
only a few callers for the exponential backoff, which are not required
to be changed.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>