pillar: containerd: Create a pipe for container's stdin
Currently the device /dev/null is used as the standard input for a
container task. However, any process reading from /dev/null will always get
0 bytes as the result, which makes interactive containers, e.g., those that
runs a shell as the entrypoint (and expects data on the standard input), to
exit right after starting. In order to keep the container task blocked (in
these situations), a valid standard input device must be provided.  This
commit creates a fake stdin that reads from an unpopulated pipe so it
blocks the reader process.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>