create standalone pkg/installer.
This is not yet consumed by anything, but will be in the next commits. The contents of pkg/installer are a combination of the following:

* install script from pkg/mkimage-raw-efi
* spec.sh from pkg/debug
* tpmmgr.go from pkg/pillar, simplified so that it only contains the commands we need for installer and verifier, and easily runs as a standalone CLI
* verification scripts from pkg/verification

There are no functional changes. However, `install` script did need some work to make it work correctly again, given that it no longer runs inside mkimage-raw-efi, and no longer can assume all of the contents where they were.

Signed-off-by: Avi Deitcher <avi@deitcher.net>