add images/installer.yml.in to build new installer image, and remove brittle references from images/rootfs.yml.in since that no longer is used for installer
We used to build installer by having the script inject containers right into an already built rootfs.yml, which requires knowing its exact order.
In short, messy and brittle, not to mention massive, as it needed the whole rootfs to boot and have the image, just to launch the installer container. This creates an installer OS image, which is dedicated to installation, has none of the brittleness or injection, and just uses installer.yml just like rootfs uses rootfs.yml

Signed-off-by: Avi Deitcher <avi@deitcher.net>