remove verification from Makefile as it no longer exists; place run-installer-net in own directory
There were multiple `make verification` type targets. With the creation of a dedicated installer image,
and all of the code from installer and verifier moved into pkg/installer, there no longer is a need for dedicated `make verification`
type targets. These all are unified under `make installer` and its variants.

In addition, when running `make run-installer-net`, it would expand the installer contents from `installer.net` tar file into `$(BUILD_DIR)/installer/`. This would overwrite anything that existed there, often breaking things needed to do a rebuild, and which were the sources of things that made `installer.net` in the first place. In order to avoid any confusion, these are moved to expand to `$(BUILD_DIR)/netboot/`, and the various Makefile commands that referenced it are updated.

Signed-off-by: Avi Deitcher <avi@deitcher.net>