update pkg/eve to work with updated artifacts for installers
No longer need to remove initrd.img and installer.img from eve/Dockerfile, since those do not exist anymore in there.
Also update how installer_net output is generated to match what is in tools/makenet.sh. Ideally, the code for both should be merged,
but there does not appear to be an easy way to do it, and it only is a few lines.

Signed-off-by: Avi Deitcher <avi@deitcher.net>