Fix bridge IP source after NI modification
There is a small bug in the master recently introduced, where
the address source of the reported bridge IP changes from
AddressSourceEVEInternal to AddressSourceUndefined when NI config
is modified in any way (note that IP source is reported only internally
between microservices, not to the controller).
While the reported source of the bridge IP is not important at all,
we should still fix it to avoid any confusion.

Signed-off-by: Milan Lenco <milan@zededa.com>