Make sure ethernet interface is DOWN before renaming and changing MAC
This is a continuation to fix 6ffa07f2f2725a5cf428a4d34a9712295623befc

On some devices, we were still encountering the error "Device or resource
busy". It turned out we had overlooked a few cases where NIM attempted
to change the MAC address while the interface was in the UP state.

Signed-off-by: Milan Lenco <milan@zededa.com>