Use DPC Reconciler to assign cluster IP to cluster interface
NIM subscribes to EdgeNodeClusterStatus published by zedkube
and passes this to DPCManager and further into DPCReconciler.
Cluster IP address is assigned to the cluster interface statically
(directly using netlink, not via dhcpcd).

zedkube subscribes to DeviceNetworkStatus to determine if the cluster
IP address is assigned and ready to use. Once assigned, it will update
EdgeNodeClusterStatus with ClusterIPIsReady set to true, which the
script cluster-init.sh is waiting for before it starts the procedure
of node joining the cluster.

Signed-off-by: Milan Lenco <milan@zededa.com>