grub: Set soft memory limits to 80% of hard limits.
By default, we now set the soft memory limits to 80% of the hard memory limits
for EVE cgroups. This adjustment sets the target values for memory reclamation
when it's triggered by the Kernel.
Updated the default values for dom0_mem, eve_mem, and ctrd_mem in the
documentation and configuration files to reflect this change.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>