zedbox: force GC on cgroups memory pressure events
Zedbox is the entry point for a Go applications in EVE, which lives
in the /eve/services cgroup, which has a memory cap of 800MB. Since
the Go runtime GC can be quite slow in its cleanup cycle, and the
system can experience OOM, this patch introduces a forced GC call
on memory pressure events.

The implementation is simple: zedbox application registers for
events from the root cgroup in hierarchy mode (events always
propagate to the root [1]), waits for mid-level memory pressure
events, and forces a GC with an interval of 10 seconds (this protects
from many cgroup events and very frequent GC calls) when certain
amount of memory has been allocated since last reclaim (same reasons
as above: to minimize unnecessary GC calls).

The put it simple: GC is called explicitly no more than once every
10 seconds and only if the application has already allocated at least
50MB and certain fraction from the last reclaim, so shortly:

   limit = MAX(50MB, reclaimed * 20%)
   expected = m.Alloc + limit

This equation should filter talkative cgroup (many sequential
memory pressure events can come) and call GC when there is something
to reclaim.

This patch is the additional measures to already introduced soft limit:
2aa31d211305 ("pillar: configure GOGC based on pillar memory limit or global config")

The main motivation is that the entire system may be under memory
pressure, so forced GC calls should help the system free up memory
even if the soft memory limit has not been reached for any Go
application.

It is worth to look for the following messages in the log:

  Received memory pressure event, before GC MemStats.Alloc=32142KB, after GC MemStats.Alloc=14590KB, reclaimed 17551KB, next GC when MemStats.Alloc=65790KB is reached

which are self explanatory.

To reproduce memory pressure events and OOM kill:

  echo $$ > /sys/fs/cgroup/memory/eve/services/tasks
  while true; do logread | grep MemStats.Alloc | wc -l; sleep 0.5; done &
  ls -d /*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*

takes up to several minutes until pressure event appears and shell
is OOM-killed.

[1] https://www.kernel.org/doc/Documentation/admin-guide/cgroup-v1/memory.rst

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>