installer: Forces shell to run interactively when required
The installer script runs within a onboot container, with output/input not
redirected to the current tty console. When a pause is required by the user
(before or after the installation process), the shell called within the
install script will not work. Forcing input/output to /dev/console fixes
the problem.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>