App-shared port without DNS server is not necessarily an issue
Not having DNS servers configured (while link is OK and IP address
is assigned), is not necessarily a failure for an app-shared interface.
Some applications might not require DNS for external hostnames
and may only use internal DNS servers to resolve the names of other apps
running on the same device. However, we should still issue a warning about
this potential issue.

Signed-off-by: Milan Lenco <milan@zededa.com>