do not try to decrypt empty wwan cipher data
Even when username/password credentials are not used for cellular
connection, zedagent still tries to decrypt the cipher block which in
this case will contain empty CipherData. This results in error
"block contains incomplete data" being reported inside the
DeviceNetworkStatus for the cellular port.  Even though EVE will not end
up publishing this to the controller (because mmagent will not try to
use CipherBlockStatus unless user auth is enabled), we should still fix
this to avoid confusing error inside the DeviceNetworkStatus pubsub message.

Signed-off-by: Milan Lenco <milan@zededa.com>