pillar/tests: Fix deadlock  when Debug Server startup fails.
In `startIntegratedPSICollectorAPI`, if the server does not start within
10 seconds, the function returns an error without unlocking
`psiServerMutex`. This oversight leads to a deadlock in subsequent tests
that attempt to acquire the same mutex.

This commit adds `psiServerMutex.Unlock()` before returning the error to
ensure that the mutex is properly released even when the server fails to
start.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>
(cherry picked from commit 7dd83ea9d01ea26b50abd387faed4de33f76dd5b)