Set WType even for wireless ports without network config
nim, zedrouter and other microservices depend on the WType port
attribute to recognize wireless network port. This is needed even in
cases where the wireless adapter is not really used and has no network
config attached. Otherwise, we may get to situation where we try to
bridge a wireless adapter or perform some other invalid operation,
especially after the switch NI enhancements done in:
7a6981ba729e0802bd1c3cb2b8da3ec2b93b8ae7

In the case wireless config is not present, we can get the physical port
type from the device model and set WType accordingly.

Signed-off-by: Milan Lenco <milan@zededa.com>