fix: update lts-release.sh to create the lts tag from the latest rc tag
feat: Add scripts and Makefile targets for RC and LTS releases

In accordance with the testing team's requirement to create a release
candidate (RC) before releasing the LTS version of Eve for certification.

These scripts automate the creation of RC and LTS releases by calculating
the version number based on the latest tag and incrementing it accordingly.

Signed-off-by: yash-zededa <yash@zededa.com>