Add volumeattach/volumedetach API
Longhorn volumes are attached to a node where the app is running.
Since the volumes we create in eve are RWO, ie only one node can attach at a time,
during failover we need to make sure the node that failed is detached from the volume.

The node that is taking over will automatically attach to the volume if no other node is
attached. Also, generally its very safe to detach and then attach to avoid any potential corruption.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>