Simplify assets.yml by using wildcards and loops
The list of assets to convert to `${{ env.ARCH }}` followed by the name, and then the list of uploads, was a fixed list. This made it break when anything changed.
This PR changes it so that it just converts all assets in `assets/` to prepend `${{ env.ARCH }}`, and uses the glob for uploading, `assets/${{ env.ARCH }}.*`

Signed-off-by: Avi Deitcher <avi@deitcher.net>