vTPM : fix bug when getting launch request
When server gets a launch request, it checks if the the requested
instance is already running, but it only checks the internal list and
not actually the running instances. This can lead to server thinking
the instance is running but client fails to get the PID with error
"failed to get pid from file ...".

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>