add support for remote log level to newlogd
As opposed to the default log level, which was setting which logs are
produced by EVE's microservices, the remote log level will set which
device logs are uploaded to the cloud. So it's assumed that the remote
log level is always higher than the default log level.

There are no changes in how newlogd collects the logs, only in how it
handles the log files.
For the logs that are uploaded:
- we create a separate file with prefix dev.log.upload in collect
- that file is gzipped when it reaches a certain size or by timer - the
  same as before
- once gzipped the file is moved to devUpload - the same as before
- once the file is uploaded successfully, it's deleted instead of being
  moved to keepSentQueue

For the logs that stay on the device:
- we create a separate file with prefix dev.log.keep in collect
- that file is gzipped when it reaches a certain size - no timer
- once gzipped the file is moved to keepSentQueue - the name of the
  folder is preserved

The commit also adds log levels "all" and "none" as well as remote log
levels for kernel and syslog logs:
as the names suggest, "all" will log everything and "none" will suppress
all logs. The levels can be set in the global configuration for
- debug.default.loglevel
- debug.default.remote.loglevel
- agent.*agentname*.debug.loglevel
- agent.*agentname*.debug.remote.loglevel
- debug.syslog.loglevel
- debug.syslog.remote.loglevel
- debug.kernel.loglevel
- debug.kernel.remote.loglevel

Signed-off-by: Paul Gaiduk <paulg@zededa.com>