Add new vTPM tests
The TestSwtpmAbruptTerminationRequest function verifies that if swtpm is
terminated without vTPM notice, no stale id is left in the vtpm internal
bookkeeping and vtpm can launch a new instance with the same id.

The TestSwtpmMultipleLaucnhRequest function verifies that if swtpm is
launched multiple times with the same id, only one instance is created
and other requests are ignored.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>