Allow to use network port with VLAN-subinterfaces for untagged traffic
With this enhancement, user will be able to use VLAN parent interface
as an L3 endpoint and access the untagged network either for EVE
management or from an application via Local network instance.

Some changes were required to allow having SystemAdapter assigned
to a network port (physical or bond) which is at the same time
referenced by a VLAN sub-interface. Previously, this was not allowed
and resulted in parsing errors published to the controller.

In terms of the Linux network configuration, not many changes were
required. We simply allow assigning static IP or running DHCP client
even for VLAN parent interface.
The only scenario which is not allowed, is to bridge network port and
at the same time use it with VLAN sub-interfaces. This means that
accessing untagged network from a Switch NI is not allowed when the port
is already used with VLAN subinterfaces.

Signed-off-by: Milan Lenco <milan@zededa.com>