remove logs longer than memlogd's max-line-len
Currently memlogd has set a max-line-len of 8192. If the log line is
longer than this, it will be truncated. Since we use structured logs
for EVE microservices, truncating logs will result in malformed JSON,
which is not parseable by newlogd.

This commit removes logs that have a potential to be longer than that
limit, because they log whole configs that tend to get very long.

In future developers should be aware of this limit and avoid logging
big chunks of data.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>