memory-monitor: Fix memory monitor config update for debug container.
Update the pkill command in the memory-monitor-update-config case to use
the process name (memory-monitor) instead of the full path
(/sbin/memory-monitor). This change ensures compatibility with the debug
container, which uses procps-ng version 3.3.17 that does not support
full paths in the pkill command.

This adjustment makes the command functional when executed in the debug
container, improving usability for users connecting via SSH.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>