Simplify mkimage-iso-efi entrypoint script and its single consumer in pkg/eve/runme.sh
The entrypoint script make-efi in mkimage-iso-efi is used in 2 places:

* the script makeiso.sh, where it expects to pipe a tar stream to the container's stdin
* the script runme.sh, as the entrypoint to pkg/eve, i.e. the container image lfedge/eve

This does not change the first use case. The make-efi is changed so that it *always* reads a tar stream from stdin.
This makes it a much simpler program to understand (consistency is nice).

For the case of runme.sh, that script is changed so that instead of assuming it has the installer in a particular directory,
it explicitly takes the installer.img, unpackages it into a temporary working directory, and creates a tar stream from the contents of that
directory, so that it can stream it to make-efi.

This has the important benefit of fixing the bug wherein making an installer_iso from `docker run lfedge/eve installer_iso` gives
an ISO that does not install correctly.

Signed-off-by: Avi Deitcher <avi@deitcher.net>
(cherry picked from commit 4df571e47f4bbf2f9457b862a7f7de3950d54036)