volumemgr: update vrs every time vs changes
Changes in volume status need to be reflected to the app. This is done
through the volume reference status. So every time the function
doUpdateVol returns we need to check if the volume status has changed
and if so update the volume reference status accordingly.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>
(cherry picked from commit 0566cd775604043ed4c88e55b916d4de75b9eaa0)