Set device to maintenance mode if TPM error is detected
We have seen errors like "point is not on the required curve" in the
past that prevents the device from upgrading because TPM fails to perform
an ECDH key agreement to generate the KEK and decrypt the vault key.
To catch these errors early, check if TPM is not functioning properly on
a critical path and set the device to maintenance mode.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>