memory-monitor: Add advanced pprof options to improve heap analysis.
Enhanced the `Makefile` to include additional options for the `go tool pprof`
command. The `-nodefraction 0` and `-call_tree` flags were added to provide
a more detailed and interactive visualization of the memory allocations,
improving the debugging and profiling process.

Updated the `README.md` to include detailed instructions for using `pprof`
effectively. This includes guidance on commands such as `top`, `list`, and
generating interactive graphs with the source code mapping.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>