Replace tty0 with tty1 when starting getty
1) tty0 points to current virtual console. it is not a problem when getty
is started very early before pillar/monitor are started however under
following conditions it causes problems:
1. there is NO 'getty' specified on kernel command line
2. there is console=tty0 specified
3. debug.enable.console=true is set

if all 3 are true then getty is started by pillar AFTER monitor app is
started on tty2 and since tty0 points to tty2 in this case getty revokes
console from monitor app and the output gets corrupted

2) we cannot use /dev/console from rungetty.sh anymore because it
corrupts tty2 when running from pillar. Instead we use /dev/tty so
the logs will be redirected to proper stdout and it should work in both
cases: when getty is started by 'getty' parameter and when it is started
from context of pillar. In later case logs will appear from pillar
context which is correct

Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>