installer: Fix root_dev() for correct INVENTORY mounting
root_dev() function in the installer script checks for the root device in
/dev/root, considering that this file will be always a symbolic link, which
is not true for some platforms, such as in some NVIDIA devices, where
/dev/root it's just a regular file block device. For these cases, INVENTORY
partition will not be found and /persist/installer will be used instead.
This commit fixes this issue by checking if /dev/root is a symbolic link or
a file block device.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>