grub: Add rootdelay for FR201 device
Some variants of FR201 device don't have an eMMC, so the whole EVE must be
installed in the NVMe disk, which is connected through an USB M.2 internal
adapter. When booting from USB storage, we need to add a rootdelay option so
kernel can wait for the USB device (during boot) in order to mount the root file
system.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>