xen-tools/init-initrd,kvm: switch shim console to an aggregator hub device
Starting from QEMU 10 (patches already backported to EVE) there is
support of an aggregator hub chardev, which aggregates input from
multiple backend devices and forwards it to a single frontend device.

For a shim console we have a local pseudo TTY (so called shim-cons,
available on `eve list-app-consoles`) and a VNC (should be explicitly
enabled by the `debug.enable.vnc.shim.vm` flag). Input from both
chardevs will be aggregated with the `hub` help and forwarded to a
virtual console `hvc1`. This gives a user possibility to control
shim VM with one virtual console from several (pty and vnc)
backends.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>